chore: refactor cleanup gov subgraph
