+++ packages/summer-earn-protocol-gov-subgraph/src/constants/index.ts
@@ -1,0 +1,5 @@
+export class EventSignature {
+  static ProposalReceivedCrossChain: string = 'ProposalReceivedCrossChain(uint256,uint32)'
+  static CallSalt: string = 'CallSalt(bytes32,bytes32)'
+  static ProposalExecuted: string = 'ProposalExecuted(uint256)'
+}
