+++ packages/summer-earn-protocol-gov-subgraph/src/mappings/timelock.ts
@@ -1,18 +1,15 @@
-import { ByteArray, Bytes, crypto, dataSource, ethereum } from '@graphprotocol/graph-ts'
-import { CrossChainProposal, CrossChainProposalByCallId } from '../../generated/schema'
+import { dataSource } from '@graphprotocol/graph-ts'
+import { CrossChainProposalByCallId } from '../../generated/schema'
 import {
   CallExecuted,
   CallSalt,
   CallScheduled,
 } from '../../generated/templates/TimelockControllerTemplate/SummerTimelockController'
-import { subgraphNetworkToChainIdMap } from './governance'
-
-export function isBase(netowrk: string): boolean {
-  return netowrk == 'base'
-}
+import { getOrCreateCrossChainProposal } from '../initializers'
+import { isHub } from '../utils/chain'
 
 export function handleCallSalt(event: CallSalt): void {
-  if (isBase(dataSource.network())) {
+  if (isHub(dataSource.network())) {
     return
   }
   const callId = event.params.id.toHexString()
@@ -25,7 +22,7 @@ export function handleCallSalt(event: CallSalt): void {
 }
 
 export function handleCallExecuted(event: CallExecuted): void {
-  if (isBase(dataSource.network())) {
+  if (isHub(dataSource.network())) {
     return
   }
   const callId = event.params.id.toHexString()
@@ -38,7 +35,7 @@ export function handleCallExecuted(event: CallExecuted): void {
 }
 
 export function handleCallScheduled(event: CallScheduled): void {
-  if (isBase(dataSource.network())) {
+  if (isHub(dataSource.network())) {
     return
   }
   const callId = event.params.id.toHexString()
@@ -60,62 +57,4 @@ export function handleCallScheduled(event: CallScheduled): void {
 
     proposal.save()
   }
-}
-
-export function getOrCreateCrossChainProposal(proposalId: string): CrossChainProposal {
-  let proposal = CrossChainProposal.load(proposalId)
-  if (!proposal) {
-    proposal = new CrossChainProposal(proposalId)
-    proposal.proposalId = proposalId
-    proposal.chainId = subgraphNetworkToChainIdMap.get(dataSource.network())
-    proposal.salt = Bytes.fromHexString('')
-    proposal.status = 'Pending'
-    proposal.targets = []
-    proposal.values = []
-    proposal.calldatas = []
-    proposal.save()
-  }
-  return proposal
-}
-
-/**
- * @dev Retrieves the event signature hash for a given event name string.
- * @param eventSig The name of the event to retrieve the signature hash for.
- * @returns A byte array representing the keccak256 hash of the event signature.
- */
-export function getTopic0(eventSig: string): ByteArray {
-  const signature = ByteArray.fromUTF8(eventSig)
-
-  return crypto.keccak256(signature)
-}
-
-export class EventSignature {
-  // EARN
-  static ProposalReceivedCrossChain: string = 'ProposalReceivedCrossChain(uint256,uint32)'
-  static CallSalt: string = 'CallSalt(bytes32,bytes32)'
-  static ProposalExecuted: string = 'ProposalExecuted(uint256)'
-}
-
-let topic0: ByteArray
-let _log: ethereum.Log
-/**
- * @dev Retrieves an array of all logs matching a particular event name.
- * @param event The Ethereum event object to retrieve logs from.
- * @param eventName The name of the event to filter for.
- * @returns An array of Ethereum log objects that match the specified event name.
- */
-export function getEventLogs(event: ethereum.Event, eventName: string): ethereum.Log[] {
-  const receipt = event.receipt
-  if (receipt == null) {
-    return []
-  }
-  topic0 = getTopic0(eventName)
-
-  return receipt.logs.filter((log) => {
-    _log = log;
-    if(_log.topics.length === 0) {
-      return false;
-    }
-    return log.topics[0].equals(topic0);
-  })
 }
